apps/frontend/AngularApp/src/app/pages/legal-doc/legal-doc-main/legal-doc-main.component.spec.ts
// testing
import { ComponentFixture } from '@angular/core/testing';
import {  configureTestingModuleForComponents, grabComponentInstance, mockTranslateService } from '@app/core/utility/test-utils';

// rxjs
import { Subject } from 'rxjs';

import { LegalDocMainComponent } from './legal-doc-main.component';


describe('LegalDocMainComponent', () => {
  let cpnt: LegalDocMainComponent;
  let fixture: ComponentFixture<LegalDocMainComponent>;

  beforeEach(async () => {
    

    
    await configureTestingModuleForComponents(LegalDocMainComponent);
    

    ({fixture, cpnt} =  grabComponentInstance(LegalDocMainComponent));
    fixture.detectChanges()
  })

  describe("init", () => {

    it("should create", () => {
      expect(cpnt).toBeTruthy()
    })

    it("should have all values initalize properly", () => {
      expect(cpnt.myClass).toEqual('LegalDocMainView')
    })

    it("should have all properties be the correct class instance", () => {
      expect(cpnt.ngUnsub).toBeInstanceOf(Subject<void>)
    })
  })

  describe("ngOnDestroy",()=>{

    beforeEach(()=>{
      spyOn(cpnt.ngUnsub,'next')
      spyOn(cpnt.ngUnsub,'complete')
    })

    it(` when called |
     as appropriate |
     does the required action `,()=>{
        // act
        cpnt.ngOnDestroy();

        // assert
        expect(cpnt.ngUnsub.next).toHaveBeenCalled();
        expect(cpnt.ngUnsub.complete).toHaveBeenCalled();
    })
  })
});
